Extract GardenaEntity base class to eliminate duplicated device_info and available
