+++ custom_components/gardena_smart_local_preview/entity.py
@@ -1,0 +1,35 @@
+from __future__ import annotations
+
+from homeassistant.helpers import device_registry as dr
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
+
+from .const import DOMAIN
+from .coordinator import GardenaSmartLocalCoordinator
+from gardena_smart_local_api.devices.device import Device
+
+
+class GardenaEntity(CoordinatorEntity[GardenaSmartLocalCoordinator]):
+    _attr_has_entity_name = True
+
+    def __init__(
+        self,
+        coordinator: GardenaSmartLocalCoordinator,
+        device: Device,
+    ) -> None:
+        super().__init__(coordinator)
+        self._device = device
+        self._attr_device_info = dr.DeviceInfo(
+            identifiers={(DOMAIN, device.id)},
+            name=f"GARDENA {device.model_definition.name} {device.serial_number}",
+            manufacturer="GARDENA",
+            model=device.model_definition.name,
+            model_id=device.model_definition.model_number,
+            sw_version=device.software_version,
+            hw_version=device.hardware_version,
+            serial_number=device.serial_number,
+        )
+
+    @property
+    def available(self) -> bool:
+        device = self.coordinator.data.get(self._device.id)
+        return bool(device and device.is_online)
