+++ custom_components/gardena_smart_local_preview/lawn_mower.py
@@ -17,12 +17,11 @@ from homeassistant.components.lawn_mower import (
 )
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import device_registry as dr
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
 from .coordinator import GardenaSmartLocalCoordinator
+from .entity import GardenaEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -53,40 +52,19 @@ async def async_setup_entry(
     coordinator.async_add_listener(_add_new_devices)
 
 
-class GardenaMower(CoordinatorEntity[GardenaSmartLocalCoordinator], LawnMowerEntity):
-    _attr_has_entity_name = True
-
+class GardenaMower(GardenaEntity, LawnMowerEntity):
     def __init__(
         self,
         coordinator: GardenaSmartLocalCoordinator,
         device: Gen1Mower1 | Gen1Mower2,
     ) -> None:
-        super().__init__(coordinator)
-        self._device = device
+        super().__init__(coordinator, device)
         self._attr_unique_id = f"{device.id}_lawn_mower"
         self._attr_name = None
         self._attr_reports_position = False
         self._attr_supported_features = (
             LawnMowerEntityFeature.DOCK | LawnMowerEntityFeature.START_MOWING
         )
-
-        self._attr_device_info = dr.DeviceInfo(
-            identifiers={(DOMAIN, device.id)},
-            name=f"GARDENA {device.model_definition.name} {device.serial_number}",
-            manufacturer=device.manufacturer,
-            model=device.model_definition.name,
-            model_id=device.model_definition.model_number,
-            sw_version=device.software_version,
-            hw_version=device.hardware_version,
-            serial_number=device.serial_number,
-        )
-
-    @property
-    def available(self) -> bool:
-        device = self.coordinator.data.get(self._device.id)
-        if not device:
-            return False
-        return device.is_online
 
     @property
     def activity(self) -> LawnMowerActivity:
