+++ custom_components/gardena_smart_local_preview/number.py
@@ -6,12 +6,11 @@ from homeassistant.components.number import NumberEntity, NumberMode
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import EntityCategory, UnitOfTime
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import device_registry as dr
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
 from .coordinator import GardenaSmartLocalCoordinator
+from .entity import GardenaEntity
 from gardena_smart_local_api.devices.device import Device
 
 _LOGGER = logging.getLogger(__name__)
@@ -45,10 +44,7 @@ async def async_setup_entry(
     coordinator.async_add_listener(_add_new_devices)
 
 
-class GardenaButtonConfigTime(
-    CoordinatorEntity[GardenaSmartLocalCoordinator], NumberEntity
-):
-    _attr_has_entity_name = True
+class GardenaButtonConfigTime(GardenaEntity, NumberEntity):
     _attr_native_min_value = 0
     _attr_native_max_value = 90
     _attr_native_step = 1
@@ -61,27 +57,9 @@ class GardenaButtonConfigTime(
         coordinator: GardenaSmartLocalCoordinator,
         device: Device,
     ) -> None:
-        super().__init__(coordinator)
-        self._device = device
+        super().__init__(coordinator, device)
         self._attr_unique_id = f"{device.id}_button_config_time"
         self._attr_name = "Button Time"
-        self._attr_device_info = dr.DeviceInfo(
-            identifiers={(DOMAIN, device.id)},
-            name=f"GARDENA {device.model_definition.name} {device.serial_number}",
-            manufacturer=device.manufacturer,
-            model=device.model_definition.name,
-            model_id=device.model_definition.model_number,
-            sw_version=device.software_version,
-            hw_version=device.hardware_version,
-            serial_number=device.serial_number,
-        )
-
-    @property
-    def available(self) -> bool:
-        device = self.coordinator.data.get(self._device.id)
-        if not device:
-            return False
-        return device.is_online
 
     @property
     def native_value(self) -> float | None:
